requirements.txt
# Telegram Bot
python-telegram-bot[job-queue]>=21.0

# Векторизация и NLP
sentence-transformers>=2.2.2
transformers>=4.35.2
torch>=2.6.0

# Векторная база данных
chromadb>=1.3.3

# Веб-интерфейс
Flask==3.0.0

# Утилиты
python-dotenv==1.0.0
requests==2.31.0
